+++ R/autoplot-Spat.R
@@ -15,6 +15,9 @@
 #'
 #' @param rgb Logical. Should be plotted as a RGB image? If `NULL` (the default)
 #'   [autoplot.SpatRaster()] would try to guess.
+#' @param use_coltab Logical. Should be plotted with the corresponding
+#'   [terra::coltab()]? If `NULL` (the default) [autoplot.SpatRaster()] would
+#'   try to guess. See also [scale_fill_coltab()].
 #' @param facets Logical. Should facets be displayed? If `NULL` (the default)
 #'   [autoplot.SpatRaster()] would try to guess.
 #' @param nrow,ncol Number of rows and columns on the facet.
@@ -59,6 +62,13 @@
 #'
 #' autoplot(tile)
 #'
+#' # With coltabs
+#'
+#' ctab <- system.file("extdata/cyl_era.tif", package = "tidyterra") %>%
+#'   rast()
+#'
+#' autoplot(ctab)
+#'
 #' #  With vectors
 #' v <- vect(system.file("extdata/cyl.gpkg", package = "tidyterra"))
 #' autoplot(v)
@@ -70,12 +80,12 @@
 autoplot.SpatRaster <- function(object,
                                 ...,
                                 rgb = NULL,
+                                use_coltab = NULL,
                                 facets = NULL,
                                 nrow = NULL, ncol = 2) {
   gg <- ggplot2::ggplot()
 
   if (is.null(rgb)) rgb <- terra::has.RGB(object)
-
 
   if (rgb) {
     gg <- gg +
@@ -85,14 +95,20 @@ autoplot.SpatRaster <- function(object,
       )
     # Done, no facets or scales on RGB
     return(gg)
-  } else {
-    gg <- gg +
-      geom_spatraster(
-        data = object,
-        ...
-      )
+  }
 
-    # Guess scale
+  # Guess scale
+  if (is.null(use_coltab)) use_coltab <- any(terra::has.colors(object))
+  gg <- gg +
+    geom_spatraster(
+      data = object,
+      use_coltab = use_coltab,
+      ...
+    )
+
+
+
+  if (!use_coltab) {
     todf <- terra::as.data.frame(object, na.rm = TRUE, xy = FALSE)
     first_lay <- unlist(lapply(todf, class))[1]