+++ R/scales_hypso.R
@@ -41,8 +41,7 @@
 #' @inheritDotParams ggplot2::discrete_scale breaks:drop
 #' @inheritDotParams ggplot2::continuous_scale breaks:labels na.value
 #' @inheritDotParams ggplot2::binned_scale breaks:limits nice.breaks
-#' @inheritParams ggplot2::scale_fill_viridis_b
-#' @inheritParams ggplot2::continuous_scale
+#' @inheritParams scale_cross_blended
 #' @param palette A valid palette name. The name is matched to the list of
 #'   available palettes, ignoring upper vs. lower case. See
 #'   [hypsometric_tints_db] for more info. Values available are:
@@ -167,7 +166,9 @@
 #'   scale_fill_hypso_tint_d(na.value = "gray10", palette = "dem_poster")
 #' }
 scale_fill_hypso_d <- function(palette = "etopo1_hypso", ...,
-                               alpha = 1, direction = 1) {
+                               alpha = 1, direction = 1,
+                               na.translate = FALSE,
+                               drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -182,13 +183,17 @@ scale_fill_hypso_d <- function(palette = "etopo1_hypso", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }
 #' @export
 #' @rdname scale_hypso
 scale_colour_hypso_d <- function(palette = "etopo1_hypso", ...,
-                                 alpha = 1, direction = 1) {
+                                 alpha = 1, direction = 1,
+                                 na.translate = FALSE,
+                                 drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -203,6 +208,8 @@ scale_colour_hypso_d <- function(palette = "etopo1_hypso", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }
@@ -386,7 +393,9 @@ hypso.colors <- function(n, palette = "etopo1_hypso",
 #' See **Examples**.
 
 scale_fill_hypso_tint_d <- function(palette = "etopo1_hypso", ...,
-                                    alpha = 1, direction = 1) {
+                                    alpha = 1, direction = 1,
+                                    na.translate = FALSE,
+                                    drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -401,13 +410,17 @@ scale_fill_hypso_tint_d <- function(palette = "etopo1_hypso", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }
 #' @export
 #' @rdname scale_hypso
 scale_colour_hypso_tint_d <- function(palette = "etopo1_hypso", ...,
-                                      alpha = 1, direction = 1) {
+                                      alpha = 1, direction = 1,
+                                      na.translate = FALSE,
+                                      drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -422,6 +435,8 @@ scale_colour_hypso_tint_d <- function(palette = "etopo1_hypso", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }