+++ R/scales_whitebox.R
@@ -29,7 +29,7 @@
 #' @inheritDotParams ggplot2::discrete_scale breaks:drop
 #' @inheritDotParams ggplot2::continuous_scale breaks:labels na.value
 #' @inheritDotParams ggplot2::binned_scale breaks:limits nice.breaks
-#' @inheritParams ggplot2::scale_fill_viridis_b
+#' @inheritParams scale_cross_blended
 #' @param palette A valid palette name. The name is matched to the list of
 #'   available palettes, ignoring upper vs. lower case. Values available are:
 #'
@@ -97,7 +97,9 @@
 #'   scale_fill_whitebox_d(na.value = "gray10", palette = "soft")
 #' }
 scale_fill_whitebox_d <- function(palette = "high_relief", ...,
-                                  alpha = 1, direction = 1) {
+                                  alpha = 1, direction = 1,
+                                  na.translate = FALSE,
+                                  drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -112,13 +114,17 @@ scale_fill_whitebox_d <- function(palette = "high_relief", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }
 #' @export
 #' @rdname scale_whitebox
 scale_colour_whitebox_d <- function(palette = "high_relief", ...,
-                                    alpha = 1, direction = 1) {
+                                    alpha = 1, direction = 1,
+                                    na.translate = FALSE,
+                                    drop = TRUE) {
   if (alpha < 0 || alpha > 1) {
     stop("alpha level ", alpha, " not in [0,1]")
   }
@@ -133,6 +139,8 @@ scale_colour_whitebox_d <- function(palette = "high_relief", ...,
       direction = direction,
       palette = palette
     ),
+    na.translate = na.translate,
+    drop = drop,
     ...
   )
 }