+++ NEWS.md
@@ -1,3 +1,12 @@
+# tidyterra (development version)
+
+-   New scales for plotting SpatRasters with color tables: `scale_fill_coltab()`
+    and rest of family scales (`colour`).
+    -   `autoplot.SpatRaster()` can detect now SpatRasters with color tables.
+    -   `geom_spatraster()` can detect now SpatRasters with color tables.
+-   By default all the discrete scales of tidyterra now have the following
+    setup: `na.translate = FALSE`.
+
 # tidyterra 0.4.0
 
 -   This release focuses heavily on SpatVectors. The improvements have been: